Complete merger of some methods
